Message context has text channel specifically
Narrowing the type to TextChannel is necessary to support coming changes
to enable the recording of messages in a session's channel.

One caveat with this however, is that TextChannel does not represent
every kind of Discord channel that is text message based, but only
general chat channels. DM and group DM channels are represented by other
classes that do not share a parent class or technically implement the
same interface.

However, it is possible they share enough of the same shape that the
type checker will allow us to later switch the type of MsgContext's
channel property to a union of all three and still call the necessary
methods without explicit casts. We'll see.