TextChannel interface
This interface abstracts the elements of the Discord.js TextChannel
class to include only the elements actually used by our code.

Thus, we can implement a much simpler test fake, which does not have to
inherit from the concrete TextChannel. This fake no longer needs to rely
on fakes for Guild, Client, etc.