Reorganize StartCommand#process tests
The diff for this looks complicated, but I only re-arranged the tests
into two sub-categories for when a session is present and when it is
not.

A couple of tests were left in the top-level "describe" block for this
function, because they should apply in both cases. However, most
behavior depends on whether there is an active session for the context's
channel or not, so this organization makes more sense.