Start command works with existing session
I had originally written the start command to do nothing if the context
already included a session. However, this is not the correct behavior,
because we want an active session to be re-startable.

Specifically, the start command should only do nothing if the recording
for the session is already active.